+++ db/migrations/0015_silent_zzzax.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "messages" ADD COLUMN "draft_product" text;